+++ interpolate_test.go
@@ -3,7 +3,7 @@ package pipeline
 import (
 	"testing"
 
-	"github.com/buildkite/go-pipeline/env"
+	"github.com/buildkite/go-pipeline/internal/env"
 	"github.com/buildkite/go-pipeline/ordered"
 	"github.com/google/go-cmp/cmp"
 	"gotest.tools/v3/assert"
